feat: First debounce and caching code
Still need to think about async to properly be true debounce. As in, the
blocked CLI calls still need to be called in the future. At the moment,
no attempt is made to recall the blocked command. You just have to wait
for the debounce to expire and make a document change again to get
latest commands run.

Also I think the debounce keys should use the cache key function that
also includes the text_doc_uri in the cache key.