fix: calculate formatter range from _current_ doc
Should fix bug where formatters (only have Black at the moment),
wouldn't sometimes misformat the end of documents. I assume this
happened because the LSP Range object, therefore the "selection" area to
replace, was based on the new incoming document, which might be longer
or shortert than what is currently in the buffer.